Create utils python module, fix bugs, add FastText embedding creation
