+++ doc_processor/processor.py
@@ -1,21 +1,14 @@
 from tensorflow import keras
-from nltk.tokenize import sent_tokenize
 from sklearn.model_selection import train_test_split
 import os
-import re
 import pickle
-import numpy as np
+import utils
 
 class DocumentCleaner:
     texts = None
     texts_dir = None
     summaries = None
     summaries_dir = None
-    text_replacements = {
-        'U. S.': 'US',
-        'U. K.': 'UK',
-        'Sen.': 'Sen'
-    }
     texts_word_dict = None
     texts_rev_word_dict = None
     text_vocab_size = None
@@ -31,31 +24,16 @@ class DocumentCleaner:
         self.texts = []
         self.summaries = []
         if text_replacements is not None:
-            self.text_replacements = text_replacements
-
-    @staticmethod
-    def is_hidden_file(name: str):
-        return name.startswith('.')
-
-    def clean_sentence(self, s: str):
-        out = s[:-1]
-        for orig, new in self.text_replacements.items():
-            out = out.replace(orig, new)
-        return out + ' <PUNCT> '
+            utils.text_cleaner.text_replacements = text_replacements
 
     def clean_texts(self):
         print(f'Ingesting and cleaning documents in {self.texts_dir}/...', end='')
         for file in os.listdir(self.texts_dir):
-            if self.is_hidden_file(file):
+            if utils.is_hidden_file(file):
                 continue
             with open(f'{self.texts_dir}/{file}', 'r', errors='ignore') as f:
                 txt = f.read()
-                # summary punctuation fix
-                fixed_txt = re.sub(r'\.(?=[^ \W\d])', '. ', txt)
-                fixed_txt = fixed_txt.replace('\n', ' <NEWLINE> ')
-                sentences = sent_tokenize(fixed_txt)
-                sentences = [self.clean_sentence(s.strip()) for s in sentences]
-                self.texts.append('<START> ' + ''.join(sentences) + '<EOS>')
+                self.texts.append(utils.clean_text(txt))
         print('done.')
 
     def tokenize_texts(self, max_length=1300, vocab_size=5000):
@@ -76,17 +54,11 @@ class DocumentCleaner:
     def clean_summaries(self):
         print(f'Ingesting and cleaning summaries in {self.summaries_dir}/...', end='')
         for file in os.listdir(self.summaries_dir):
-            if self.is_hidden_file(file):
+            if utils.is_hidden_file(file):
                 continue
             with open(f'{self.summaries_dir}/{file}', 'r', errors='ignore') as f:
                 txt = f.read()
-                # summary punctuation fix
-                fixed_txt = re.sub(r'\.(?=[^ \W\d])', '. ', txt)
-                fixed_txt = fixed_txt.replace('\n', ' <NEWLINE> ')
-                fixed_txt = fixed_txt.replace('\xa0', ' ')
-                sentences = sent_tokenize(fixed_txt)
-                sentences = [self.clean_sentence(s.strip()) for s in sentences]
-                self.summaries.append('<START> ' + ''.join(sentences) + '<EOS>')
+                self.summaries.append(utils.clean_text(txt))
         print('done.')
 
     def tokenize_summaries(self, max_length=150, vocab_size=5000):
@@ -98,6 +70,7 @@ class DocumentCleaner:
         self.summaries = keras.preprocessing.sequence.pad_sequences(seq, maxlen=max_length, truncating='post')
         word_dict = tokenizer.word_index
         word_dict['<pad>'] = 0
+        tokenizer.index_word[0] = '<pad>'
         self.summaries_word_dict = word_dict
         self.summaries_rev_word_dict = dict([(value, key) for (key, value) in self.summaries_word_dict.items()])
         self.summary_tokenizer = tokenizer
