+++ utils/__init__.py
@@ -1,0 +1,2 @@
+from .text_cleaner import is_hidden_file
+from .text_cleaner import clean_text
